A frame from the race used to generate data
Only the top right corner including the table is relevant for us, since it contains the leads of all opponents. This table is separated into four rows. Each individual row is then separated into five "boxes". Each box belongs to one of the following classes: 0,1,2,3,4,5,6,7,8,9,+,-,void,ewil,name . 'void' is used to show that this box does not contain information. 'ewil' is the name of the player in our case, and 'name' is used for boxes containing letters of the names of opponents.